+++ src/library/DIPS.Mobile.UI/Components/ContextMenus/GlobalContextMenuClickMetadata.cs
@@ -1,0 +1,3 @@
+namespace DIPS.Mobile.UI.Components.ContextMenus;
+
+public record GlobalContextMenuClickMetadata(ContextMenuItem ContextMenuItem, ContextMenu ContextMenu);