+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## [41.0.0]
+- [ContextMenu] `SetContextMenuItemClickedCallback` is renamed to `ItemsShouldSendGlobalClicks`.
+- [ContextMenu] `ItemsShouldSendGlobalClicks` now sends a `GlobalContextMenuClickMetadata`.
+- [ContextMenu] Added `Mode` getter to know which mode a menu has.
+- [ContextMenuItem] Added `ShouldSendGlobalClick` which controls if all Context Menu item will send a logging callback. Default false.
+
 ## [40.3.9]
 - [Entry][iOS] Added 'Done' toolbar item. 
 - [Entry][iOS] Attempt to avoid seemingly random crash.